+++ Bungeecord/src/main/java/de/btegermany/teleportation/TeleportationBungee/registry/RegistriesProvider.java
@@ -9,12 +9,16 @@ public class RegistriesProvider {
     private final TpasRegistry tpasRegistry;
     private final BukkitPlayersRegistry bukkitPlayersRegistry;
     private final WarpsRegistry warpsRegistry;
+    private final SentCoordinatesFormatWarningRegistry sentCoordinatesFormatWarningRegistry;
+    private final WarpTagsRegistry warpTagsRegistry;
 
     public RegistriesProvider(Database database, TeleportationBungee plugin) {
         this.lastLocationsRegistry = new LastLocationsRegistry();
         this.tpasRegistry = new TpasRegistry();
         this.bukkitPlayersRegistry = new BukkitPlayersRegistry();
-        this.warpsRegistry = new WarpsRegistry(database, plugin);
+        this.warpTagsRegistry = new WarpTagsRegistry();
+        this.warpsRegistry = new WarpsRegistry(database, plugin, warpTagsRegistry);
+        this.sentCoordinatesFormatWarningRegistry = new SentCoordinatesFormatWarningRegistry();
     }
 
     public LastLocationsRegistry getLastLocationsRegistry() {
@@ -31,5 +35,13 @@ public class RegistriesProvider {
 
     public WarpsRegistry getWarpsRegistry() {
         return warpsRegistry;
+    }
+
+    public SentCoordinatesFormatWarningRegistry getSentCoordinatesFormatWarningRegistry() {
+        return sentCoordinatesFormatWarningRegistry;
+    }
+
+    public WarpTagsRegistry getWarpTagsRegistry() {
+        return warpTagsRegistry;
     }
 }